BR-42 : Integrate API for next-appointment
